Moved RenameMap to new file. Deprecated RenameMap.apply
